Renamed contactType to contactRecipient and moved UserData to parent folder
